+++ apps/taalwiz-api/src/hashtag/hashtag.controller.ts
@@ -1,4 +1,6 @@
-import { Controller, Get, Param } from '@nestjs/common';
+import { Controller, Get, Param, Req } from '@nestjs/common';
+import type { Request } from 'express';
+import type { JwtPayload } from '../auth/types/jwtpayload.interface.js';
 import { HashtagService } from './hashtag.service.js';
 
 @Controller('hashtags')
@@ -6,12 +8,12 @@ export class HashtagController {
   constructor(private readonly hashtagService: HashtagService) {}
 
   @Get()
-  async getHashtagIndex() {
-    return await this.hashtagService.getHashtagIndex();
+  async getHashtagIndex(@Req() req: Request) {
+    return await this.hashtagService.getHashtagIndex(req['user'] as JwtPayload);
   }
 
   @Get(':name')
-  async findHashtag(@Param('name') name: string) {
-    return await this.hashtagService.findHashtag(name);
+  async findHashtag(@Param('name') name: string, @Req() req: Request) {
+    return await this.hashtagService.findHashtag(name, req['user'] as JwtPayload);
   }
 }