+++ apps/taalwiz-api/src/users/models/user.model.ts
@@ -8,6 +8,7 @@ const UserSchema = new Schema({
   email: { type: String, lowercase: true, required: true, unique: true },
   password: { type: String, required: true },
   roles: { type: [String], default: ['user'], enum: ['user', 'admin', 'demo'] },
+  groups: { type: [String], default: [] },
   lang: { type: String, default: 'nl', enum: ['en', 'nl'] },
   created: { type: Date, default: Date.now },
   lastAccessed: { type: Date, default: Date.now },